notifierとproviderのファイルを分離
